Refactor advanced ML models into dedicated modules
