+++ models/ml_models.py
@@ -1,1298 +1,11 @@
-import logging
-from datetime import datetime
-from typing import Any, Dict, List, Optional, Tuple
-
-import numpy as np
-import pandas as pd
-import torch
-import torch.nn as nn
-import torch.optim as optim
-
-from data.stock_data import StockDataProvider
-
-from .cache import AdvancedCacheManager, RedisCache
-from .deep_learning import DeepLearningPredictor, DQNReinforcementLearner
-from .ensemble_predictor import (
-    AdvancedEnsemblePredictor,
-    EnsembleStockPredictor,
-    ParallelStockPredictor,
-)
-from .meta_learning import MetaLearningOptimizer
-from .ml_stock_predictor import MLStockPredictor
-from .sentiment import MacroEconomicDataProvider, SentimentAnalyzer
-
-logger = logging.getLogger(__name__)
-
-class AdvancedPrecisionBreakthrough87System:
-    """
-    87%精度突破システム
-    5つのブレークスルー技術:
-    1. 強化学習 (Deep Q-Network)
-    2. マルチモーダル分析 (CNN + LSTM)
-    3. メタ学習最適化 (MAML)
-    4. 高度アンサンブル最適化
-    5. 時系列Transformer最適化
-    """
-
-    def __init__(self):
-        self.logger = logging.getLogger(__name__)
-        self.models = {}
-        self.ensemble_weights = {}
-        self.current_accuracy = 84.6
-        self.target_accuracy = 87.0
-        # 各コンポーネント初期化
-        self._initialize_components()
-
-    def _initialize_components(self):
-        """コンポーネント初期化"""
-        try:
-            self.logger.info("87%精度突破システム初期化開始")
-            # 1. 強化学習エージェント
-            self.dqn_agent = self._create_dqn_agent()
-            # 2. マルチモーダル分析器
-            self.multimodal_analyzer = self._create_multimodal_analyzer()
-            # 3. メタ学習オプティマイザー
-            self.meta_optimizer = self._create_meta_optimizer()
-            # 4. 高度アンサンブル
-            self.advanced_ensemble = self._create_advanced_ensemble()
-            # 5. 時系列Transformer
-            self.market_transformer = self._create_market_transformer()
-            self.logger.info("87%精度突破システム初期化完了")
-        except Exception as e:
-            self.logger.error(f"初期化エラー: {e}")
-
-    def _create_dqn_agent(self):
-        """Deep Q-Network強化学習エージェント作成"""
-        try:
-            import random
-            from collections import deque
-
-            class DQNNetwork(nn.Module):
-                def __init__(self, state_size=50, action_size=3, hidden_size=256):
-                    super(DQNNetwork, self).__init__()
-                    self.fc1 = nn.Linear(state_size, hidden_size)
-                    self.fc2 = nn.Linear(hidden_size, hidden_size)
-                    self.fc3 = nn.Linear(hidden_size, hidden_size)
-                    self.fc4 = nn.Linear(hidden_size, action_size)
-                    self.dropout = nn.Dropout(0.3)
-
-                def forward(self, x):
-                    x = torch.relu(self.fc1(x))
-                    x = self.dropout(x)
-                    x = torch.relu(self.fc2(x))
-                    x = self.dropout(x)
-                    x = torch.relu(self.fc3(x))
-                    return self.fc4(x)
-
-            class StockTradingDQN:
-                def __init__(self):
-                    self.state_size = 50
-                    self.action_size = 3  # 買い/売り/ホールド
-                    self.memory = deque(maxlen=10000)
-                    self.epsilon = 1.0
-                    self.epsilon_min = 0.01
-                    self.epsilon_decay = 0.995
-                    self.learning_rate = 0.001
-                    # ニューラルネットワーク
-                    self.q_network = DQNNetwork()
-                    self.target_network = DQNNetwork()
-                    self.optimizer = optim.Adam(
-                        self.q_network.parameters(), lr=self.learning_rate
-                    )
-
-                def remember(self, state, action, reward, next_state, done):
-                    """経験を記憶"""
-                    self.memory.append((state, action, reward, next_state, done))
-
-                def act(self, state):
-                    """行動選択 (ε-greedy)"""
-                    if np.random.random() <= self.epsilon:
-                        return random.randrange(self.action_size)  # nosec B311
-                    state_tensor = torch.FloatTensor(state).unsqueeze(0)
-                    q_values = self.q_network(state_tensor)
-                    return q_values.argmax().item()
-
-                def replay(self, batch_size=32):
-                    """経験リプレイ学習"""
-                    if len(self.memory) < batch_size:
-                        return
-                    batch = random.sample(self.memory, batch_size)  # nosec B311
-                    states = torch.FloatTensor([e[0] for e in batch])
-                    actions = torch.LongTensor([e[1] for e in batch])
-                    rewards = torch.FloatTensor([e[2] for e in batch])
-                    next_states = torch.FloatTensor([e[3] for e in batch])
-                    dones = torch.BoolTensor([e[4] for e in batch])
-                    current_q_values = self.q_network(states).gather(
-                        1, actions.unsqueeze(1)
-                    )
-                    next_q_values = self.target_network(next_states).max(1)[0].detach()
-                    target_q_values = rewards + (0.95 * next_q_values * ~dones)
-                    loss = nn.MSELoss()(current_q_values.squeeze(), target_q_values)
-                    self.optimizer.zero_grad()
-                    loss.backward()
-                    self.optimizer.step()
-                    if self.epsilon > self.epsilon_min:
-                        self.epsilon *= self.epsilon_decay
-
-                def update_target_network(self):
-                    """ターゲットネットワーク更新"""
-                    self.target_network.load_state_dict(self.q_network.state_dict())
-
-                def predict_with_dqn(self, market_state):
-                    """DQN予測実行"""
-                    state_tensor = torch.FloatTensor(market_state).unsqueeze(0)
-                    with torch.no_grad():
-                        q_values = self.q_network(state_tensor)
-                        confidence = torch.softmax(q_values, dim=1).max().item()
-                    return {
-                        "action": q_values.argmax().item(),
-                        "confidence": confidence,
-                        "q_values": q_values.numpy(),
-                    }
-
-            return StockTradingDQN()
-        except ImportError:
-            self.logger.warning("PyTorch不可 - DQN簡易版使用")
-            return self._create_simple_dqn()
-        except Exception as e:
-            self.logger.error(f"DQNエージェント作成エラー: {e}")
-            return None
-
-    def _create_simple_dqn(self):
-        """簡易DQNエージェント"""
-
-        class SimpleDQN:
-            def predict_with_dqn(self, market_state):
-                # 簡易版: 移動平均ベース判断
-                momentum = np.mean(market_state[-5:]) - np.mean(market_state[-10:-5])
-                if momentum > 0.01:
-                    action = 0  # 買い
-                elif momentum < -0.01:
-                    action = 1  # 売り
-                else:
-                    action = 2  # ホールド
-                confidence = min(abs(momentum) * 10, 1.0)
-                return {"action": action, "confidence": confidence}
-
-        return SimpleDQN()
-
-    def _create_multimodal_analyzer(self):
-        """マルチモーダル分析器作成"""
-        try:
-            import cv2
-            from PIL import Image, ImageDraw
-            import matplotlib.pyplot as plt
-            import io
-
-            class MultiModalAnalyzer:
-                def __init__(self):
-                    self.cnn_features_size = 128
-                    self.lstm_features_size = 64
-
-                def create_chart_image(self, price_data):
-                    """価格データからチャート画像作成"""
-                    try:
-                        plt.figure(figsize=(8, 6))
-                        plt.plot(price_data, linewidth=2)
-                        plt.title("Price Chart")
-                        plt.grid(True, alpha=0.3)
-                        # 画像をバイトデータに変換
-                        img_buffer = io.BytesIO()
-                        plt.savefig(
-                            img_buffer, format="png", dpi=100, bbox_inches="tight"
-                        )
-                        img_buffer.seek(0)
-                        # PILで読み込み
-                        img = Image.open(img_buffer)
-                        img_array = np.array(img)
-                        plt.close()
-                        return img_array
-                    except Exception as e:
-                        # エラー時は簡易パターン作成
-                        return np.random.randint(0, 255, (100, 100, 3), dtype=np.uint8)
-
-                def extract_chart_features(self, chart_image):
-                    """チャート画像から特徴量抽出"""
-                    try:
-                        # 簡易版: 画像の統計的特徴量
-                        if len(chart_image.shape) == 3:
-                            gray = np.mean(chart_image, axis=2)
-                        else:
-                            gray = chart_image
-                        # 特徴量計算
-                        features = [
-                            np.mean(gray),  # 平均輝度
-                            np.std(gray),  # 輝度標準偏差
-                            np.max(gray),  # 最大輝度
-                            np.min(gray),  # 最小輝度
-                            len(np.where(np.diff(gray.flatten()) > 5)[0]),  # エッジ数
-                        ]
-                        # 128次元に拡張（ゼロパディング）
-                        features.extend(
-                            [0.0] * (self.cnn_features_size - len(features))
-                        )
-                        return np.array(features[: self.cnn_features_size])
-                    except Exception as e:
-                        return np.zeros(self.cnn_features_size)
-
-                def extract_numerical_features(self, time_series_data):
-                    """数値時系列データから特徴量抽出"""
-                    try:
-                        # LSTM風特徴量
-                        if len(time_series_data) < 10:
-                            return np.zeros(self.lstm_features_size)
-                        # 時系列統計特徴量
-                        features = [
-                            np.mean(time_series_data),
-                            np.std(time_series_data),
-                            np.max(time_series_data),
-                            np.min(time_series_data),
-                            np.mean(np.diff(time_series_data)),  # 平均変化率
-                            np.std(np.diff(time_series_data)),  # 変化率標準偏差
-                        ]
-                        # 移動平均特徴量
-                        for window in [5, 10, 20]:
-                            if len(time_series_data) >= window:
-                                ma = np.mean(time_series_data[-window:])
-                                features.append(ma)
-                                features.append(time_series_data[-1] - ma)  # 乖離
-                        # 64次元に調整
-                        features.extend(
-                            [0.0] * (self.lstm_features_size - len(features))
-                        )
-                        return np.array(features[: self.lstm_features_size])
-                    except Exception as e:
-                        return np.zeros(self.lstm_features_size)
-
-                def fuse_features(self, chart_features, numerical_features):
-                    """特徴量融合"""
-                    try:
-                        # 重み付き結合
-                        chart_weight = 0.4
-                        numerical_weight = 0.6
-                        # 正規化
-                        chart_norm = chart_features / (
-                            np.linalg.norm(chart_features) + 1e-8
-                        )
-                        numerical_norm = numerical_features / (
-                            np.linalg.norm(numerical_features) + 1e-8
-                        )
-                        # 融合
-                        fused = np.concatenate(
-                            [
-                                chart_norm * chart_weight,
-                                numerical_norm * numerical_weight,
-                            ]
-                        )
-                        return fused
-                    except Exception as e:
-                        return np.zeros(
-                            self.cnn_features_size + self.lstm_features_size
-                        )
-
-                def predict_multimodal(self, price_data, volume_data=None):
-                    """マルチモーダル予測"""
-                    try:
-                        # チャート画像作成・特徴量抽出
-                        chart_image = self.create_chart_image(price_data)
-                        chart_features = self.extract_chart_features(chart_image)
-                        # 数値特徴量抽出
-                        numerical_features = self.extract_numerical_features(price_data)
-                        # 特徴量融合
-                        fused_features = self.fuse_features(
-                            chart_features, numerical_features
-                        )
-                        # 簡易予測（融合特徴量の線形結合）
-                        prediction_score = np.mean(fused_features) * 100
-                        # 信頼度計算
-                        confidence = min(np.std(fused_features) * 2, 1.0)
-                        return {
-                            "prediction_score": prediction_score,
-                            "confidence": confidence,
-                            "chart_features": chart_features,
-                            "numerical_features": numerical_features,
-                            "fused_features": fused_features,
-                        }
-                    except Exception as e:
-                        return {
-                            "prediction_score": 50.0,
-                            "confidence": 0.0,
-                            "error": str(e),
-                        }
-
-            return MultiModalAnalyzer()
-        except ImportError as e:
-            self.logger.warning(f"マルチモーダル依存関係不足: {e}")
-            return self._create_simple_multimodal()
-        except Exception as e:
-            self.logger.error(f"マルチモーダル分析器作成エラー: {e}")
-            return None
-
-    def _create_simple_multimodal(self):
-        """簡易マルチモーダル分析器"""
-
-        class SimpleMultiModal:
-            def predict_multimodal(self, price_data, volume_data=None):
-                # 簡易版: 価格統計ベース
-                trend = np.mean(price_data[-5:]) - np.mean(price_data[-10:-5])
-                volatility = np.std(price_data[-20:])
-                score = 50 + trend * 1000 + (0.1 - volatility) * 100
-                confidence = min(abs(trend) * 100, 1.0)
-                return {
-                    "prediction_score": max(0, min(100, score)),
-                    "confidence": confidence,
-                }
-
-        return SimpleMultiModal()
-
-    def _create_meta_optimizer(self):
-        """メタ学習オプティマイザー作成"""
-
-        class MetaLearningOptimizer:
-            def __init__(self):
-                self.symbol_adaptations = {}
-                self.sector_patterns = {}
-
-            def adapt_to_symbol(self, symbol, historical_performance):
-                """銘柄特性に適応"""
-                try:
-                    # 銘柄固有パターン学習
-                    if len(historical_performance) >= 10:
-                        avg_performance = np.mean(historical_performance)
-                        volatility = np.std(historical_performance)
-                        trend = np.polyfit(
-                            range(len(historical_performance)),
-                            historical_performance,
-                            1,
-                        )[0]
-                        adaptation = {
-                            "performance_bias": avg_performance - 50,
-                            "volatility_factor": volatility,
-                            "trend_factor": trend,
-                            "adaptation_strength": min(
-                                len(historical_performance) / 50, 1.0
-                            ),
-                        }
-                        self.symbol_adaptations[symbol] = adaptation
-                        return adaptation
-                    return {"adaptation_strength": 0.0}
-                except Exception as e:
-                    return {"adaptation_strength": 0.0, "error": str(e)}
-
-            def get_sector_adaptation(self, symbol):
-                """セクター別適応"""
-                try:
-                    # 簡易セクター判定
-                    first_digit = symbol[0] if symbol else "0"
-                    sector_map = {
-                        "1": "construction",
-                        "2": "foods",
-                        "3": "textiles",
-                        "4": "chemicals",
-                        "5": "pharmacy",
-                        "6": "metals",
-                        "7": "machinery",
-                        "8": "electronics",
-                        "9": "transport",
-                    }
-                    sector = sector_map.get(first_digit, "others")
-                    # セクター固有調整
-                    sector_adjustments = {
-                        "electronics": {
-                            "volatility_multiplier": 1.2,
-                            "trend_sensitivity": 1.1,
-                        },
-                        "machinery": {
-                            "volatility_multiplier": 0.9,
-                            "trend_sensitivity": 1.0,
-                        },
-                        "transport": {
-                            "volatility_multiplier": 1.1,
-                            "trend_sensitivity": 0.9,
-                        },
-                        "others": {
-                            "volatility_multiplier": 1.0,
-                            "trend_sensitivity": 1.0,
-                        },
-                    }
-                    return sector_adjustments.get(sector, sector_adjustments["others"])
-                except Exception as e:
-                    return {"volatility_multiplier": 1.0, "trend_sensitivity": 1.0}
-
-            def meta_predict(self, symbol, base_prediction):
-                """メタ学習予測"""
-                try:
-                    # 銘柄適応
-                    symbol_adaptation = self.symbol_adaptations.get(
-                        symbol, {"adaptation_strength": 0.0}
-                    )
-                    # セクター適応
-                    sector_adaptation = self.get_sector_adaptation(symbol)
-                    # 適応強度に応じた調整
-                    adaptation_strength = symbol_adaptation.get(
-                        "adaptation_strength", 0.0
-                    )
-                    if adaptation_strength > 0.1:
-                        # 適応調整適用
-                        bias = symbol_adaptation.get("performance_bias", 0)
-                        trend_factor = symbol_adaptation.get("trend_factor", 0)
-                        adjusted_prediction = (
-                            base_prediction + bias * adaptation_strength
-                        )
-                        adjusted_prediction += trend_factor * 10 * adaptation_strength
-                        # セクター調整
-                        volatility_mult = sector_adaptation.get(
-                            "volatility_multiplier", 1.0
-                        )
-                        adjusted_prediction = (
-                            50 + (adjusted_prediction - 50) * volatility_mult
-                        )
-                        confidence_boost = adaptation_strength * 0.1
-                        return {
-                            "adjusted_prediction": max(
-                                0, min(100, adjusted_prediction)
-                            ),
-                            "confidence_boost": confidence_boost,
-                            "adaptation_applied": True,
-                        }
-                    return {
-                        "adjusted_prediction": base_prediction,
-                        "confidence_boost": 0.0,
-                        "adaptation_applied": False,
-                    }
-                except Exception as e:
-                    return {
-                        "adjusted_prediction": base_prediction,
-                        "confidence_boost": 0.0,
-                        "adaptation_applied": False,
-                        "error": str(e),
-                    }
-
-        return MetaLearningOptimizer()
-
-    def _create_advanced_ensemble(self):
-        """高度アンサンブル作成"""
-
-        class AdvancedEnsemble:
-            def __init__(self):
-                self.base_weights = {
-                    "trend_following": 0.35,  # 84.6%ベース重視
-                    "dqn": 0.20,
-                    "multimodal": 0.20,
-                    "meta": 0.15,
-                    "transformer": 0.10,
-                }
-                self.performance_history = {}
-
-            def update_weights_dynamically(self, recent_performances):
-                """動的重み調整"""
-                try:
-                    if not recent_performances:
-                        return self.base_weights
-                    # パフォーマンスベース重み調整
-                    total_performance = sum(recent_performances.values())
-                    if total_performance > 0:
-                        adjusted_weights = {}
-                        for model, base_weight in self.base_weights.items():
-                            performance = recent_performances.get(model, base_weight)
-                            adjusted_weight = base_weight * (
-                                1 + (performance - 0.5) * 0.3
-                            )
-                            adjusted_weights[model] = max(
-                                0.05, min(0.6, adjusted_weight)
-                            )
-                        # 正規化
-                        total_weight = sum(adjusted_weights.values())
-                        if total_weight > 0:
-                            adjusted_weights = {
-                                k: v / total_weight for k, v in adjusted_weights.items()
-                            }
-                            return adjusted_weights
-                    return self.base_weights
-                except Exception as e:
-                    return self.base_weights
-
-            def ensemble_predict(self, predictions, confidences):
-                """アンサンブル予測実行"""
-                try:
-                    # 動的重み取得
-                    weights = self.update_weights_dynamically(confidences)
-                    # 信頼度重み付きアンサンブル
-                    weighted_sum = 0.0
-                    total_weight = 0.0
-                    for model, prediction in predictions.items():
-                        if model in weights and prediction is not None:
-                            confidence = confidences.get(model, 0.5)
-                            model_weight = weights[model]
-                            # 信頼度 × モデル重み
-                            effective_weight = model_weight * (0.5 + confidence * 0.5)
-                            weighted_sum += prediction * effective_weight
-                            total_weight += effective_weight
-                    if total_weight > 0:
-                        ensemble_prediction = weighted_sum / total_weight
-                        ensemble_confidence = min(
-                            total_weight / sum(weights.values()), 1.0
-                        )
-                    else:
-                        ensemble_prediction = 50.0
-                        ensemble_confidence = 0.0
-                    return {
-                        "ensemble_prediction": ensemble_prediction,
-                        "ensemble_confidence": ensemble_confidence,
-                        "used_weights": weights,
-                        "total_weight": total_weight,
-                    }
-                except Exception as e:
-                    return {
-                        "ensemble_prediction": 50.0,
-                        "ensemble_confidence": 0.0,
-                        "error": str(e),
-                    }
-
-        return AdvancedEnsemble()
-
-    def _create_market_transformer(self):
-        """市場専用Transformer作成"""
-
-        class MarketTransformer:
-            def __init__(self):
-                self.sequence_length = 60
-                self.feature_dim = 10
-
-            def create_market_features(self, price_data, volume_data=None):
-                """市場特徴量作成"""
-                try:
-                    if len(price_data) < self.sequence_length:
-                        # データ不足時はゼロパディング
-                        padded_data = np.zeros(self.sequence_length)
-                        padded_data[-len(price_data) :] = price_data
-                        price_data = padded_data
-                    # 時系列特徴量
-                    features = []
-                    for i in range(len(price_data) - self.sequence_length + 1):
-                        window = price_data[i : i + self.sequence_length]
-                        # 基本統計
-                        feature_vector = [
-                            np.mean(window),
-                            np.std(window),
-                            np.max(window),
-                            np.min(window),
-                            window[-1] - window[0],  # 変化量
-                            np.mean(np.diff(window)),  # 平均変化率
-                            len(np.where(np.diff(window) > 0)[0])
-                            / len(window),  # 上昇率
-                        ]
-                        # 10次元に調整
-                        feature_vector.extend(
-                            [0.0] * (self.feature_dim - len(feature_vector))
-                        )
-                        features.append(feature_vector[: self.feature_dim])
-                    return (
-                        np.array(features)
-                        if features
-                        else np.zeros((1, self.feature_dim))
-                    )
-                except Exception as e:
-                    return np.zeros((1, self.feature_dim))
-
-            def transformer_attention(self, features):
-                """簡易アテンション機構"""
-                try:
-                    if len(features.shape) != 2:
-                        return np.mean(features)
-                    # 簡易セルフアテンション
-                    attention_weights = np.exp(np.sum(features, axis=1))
-                    attention_weights = attention_weights / np.sum(attention_weights)
-                    # 重み付き平均
-                    attended_features = np.average(
-                        features, axis=0, weights=attention_weights
-                    )
-                    return attended_features
-                except Exception as e:
-                    return (
-                        np.mean(features, axis=0)
-                        if len(features.shape) == 2
-                        else np.zeros(self.feature_dim)
-                    )
-
-            def transformer_predict(self, price_data, volume_data=None):
-                """Transformer予測"""
-                try:
-                    # 市場特徴量作成
-                    features = self.create_market_features(price_data, volume_data)
-                    # アテンション適用
-                    attended = self.transformer_attention(features)
-                    # 予測計算（簡易版）
-                    prediction_score = 50 + np.sum(attended) * 5
-                    prediction_score = max(0, min(100, prediction_score))
-                    # 信頼度計算
-                    confidence = min(np.std(attended) * 0.5, 1.0)
-                    return {
-                        "prediction_score": prediction_score,
-                        "confidence": confidence,
-                        "attention_weights": attended,
-                    }
-                except Exception as e:
-                    return {
-                        "prediction_score": 50.0,
-                        "confidence": 0.0,
-                        "error": str(e),
-                    }
-
-        return MarketTransformer()
-
-    def predict_87_percent_accuracy(self, symbol: str) -> Dict[str, Any]:
-        """87%精度予測実行"""
-        try:
-            self.logger.info(f"87%精度予測開始: {symbol}")
-            # データ取得
-            price_data, volume_data = self._get_market_data(symbol)
-            if price_data is None or len(price_data) < 20:
-                return self._return_fallback_prediction(symbol)
-            # 各モデルで予測実行
-            predictions = {}
-            confidences = {}
-            # 1. ベース予測（84.6%システム）
-            base_result = self._get_base_prediction(symbol, price_data)
-            predictions["trend_following"] = base_result["prediction"]
-            confidences["trend_following"] = base_result["confidence"]
-            # 2. DQN予測
-            if self.dqn_agent:
-                market_state = self._create_market_state(price_data, volume_data)
-                dqn_result = self.dqn_agent.predict_with_dqn(market_state)
-                predictions["dqn"] = self._convert_action_to_score(dqn_result["action"])
-                confidences["dqn"] = dqn_result["confidence"]
-            # 3. マルチモーダル予測
-            if self.multimodal_analyzer:
-                multimodal_result = self.multimodal_analyzer.predict_multimodal(
-                    price_data, volume_data
-                )
-                predictions["multimodal"] = multimodal_result["prediction_score"]
-                confidences["multimodal"] = multimodal_result["confidence"]
-            # 4. メタ学習予測
-            if self.meta_optimizer and "trend_following" in predictions:
-                meta_result = self.meta_optimizer.meta_predict(
-                    symbol, predictions["trend_following"]
-                )
-                predictions["meta"] = meta_result["adjusted_prediction"]
-                confidences["meta"] = (
-                    base_result["confidence"] + meta_result["confidence_boost"]
-                )
-            # 5. Transformer予測
-            if self.market_transformer:
-                transformer_result = self.market_transformer.transformer_predict(
-                    price_data, volume_data
-                )
-                predictions["transformer"] = transformer_result["prediction_score"]
-                confidences["transformer"] = transformer_result["confidence"]
-            # 高度アンサンブル実行
-            ensemble_result = self.advanced_ensemble.ensemble_predict(
-                predictions, confidences
-            )
-            # 87%精度補正
-            final_prediction = self._apply_87_percent_correction(
-                ensemble_result["ensemble_prediction"],
-                ensemble_result["ensemble_confidence"],
-                symbol,
-            )
-            result = {
-                "symbol": symbol,
-                "final_prediction": final_prediction["prediction"],
-                "final_confidence": final_prediction["confidence"],
-                "target_accuracy": 87.0,
-                "individual_predictions": predictions,
-                "individual_confidences": confidences,
-                "ensemble_result": ensemble_result,
-                "accuracy_improvement": final_prediction["prediction"]
-                - self.current_accuracy,
-                "model_contributions": self._analyze_model_contributions(
-                    predictions, confidences
-                ),
-            }
-            self.logger.info(
-                f"87%精度予測完了: {symbol}, 予測={final_prediction['prediction']:.1f}"
-            )
-            return result
-        except Exception as e:
-            self.logger.error(f"87%精度予測エラー {symbol}: {e}")
-            return self._return_fallback_prediction(symbol, error=str(e))
-
-    def _get_market_data(self, symbol: str) -> Tuple[np.ndarray, np.ndarray]:
-        """市場データ取得"""
-        try:
-            import yfinance as yf
-
-            ticker = yf.Ticker(f"{symbol}.T")
-            data = ticker.history(period="1y")
-            if data.empty:
-                return None, None
-            price_data = data["Close"].values
-            volume_data = data["Volume"].values if "Volume" in data else None
-            return price_data, volume_data
-        except Exception as e:
-            self.logger.warning(f"市場データ取得エラー {symbol}: {e}")
-            return None, None
-
-    def _get_base_prediction(
-        self, symbol: str, price_data: np.ndarray
-    ) -> Dict[str, float]:
-        """84.6%ベース予測取得"""
-        try:
-            # 簡易トレンドフォロー
-            if len(price_data) >= 50:
-                sma_10 = np.mean(price_data[-10:])
-                sma_20 = np.mean(price_data[-20:])
-                sma_50 = np.mean(price_data[-50:])
-                # トレンド判定
-                trend_bullish = sma_10 > sma_20 > sma_50
-                trend_bearish = sma_10 < sma_20 < sma_50
-                if trend_bullish:
-                    prediction = 75.0
-                    confidence = 0.8
-                elif trend_bearish:
-                    prediction = 25.0
-                    confidence = 0.8
-                else:
-                    prediction = 50.0
-                    confidence = 0.5
-                return {"prediction": prediction, "confidence": confidence}
-            return {"prediction": 50.0, "confidence": 0.3}
-        except Exception as e:
-            return {"prediction": 50.0, "confidence": 0.0}
-
-    def _create_market_state(
-        self, price_data: np.ndarray, volume_data: np.ndarray
-    ) -> np.ndarray:
-        """DQN用市場状態作成"""
-        try:
-            state_size = 50
-            if len(price_data) < state_size:
-                # データ不足時はゼロパディング
-                state = np.zeros(state_size)
-                state[-len(price_data) :] = price_data[-len(price_data) :]
-            else:
-                # 正規化した価格データ
-                recent_prices = price_data[-state_size:]
-                state = (recent_prices - np.mean(recent_prices)) / (
-                    np.std(recent_prices) + 1e-8
-                )
-            return state
-        except Exception as e:
-            return np.zeros(50)
-
-    def _convert_action_to_score(self, action: int) -> float:
-        """DQNアクション → スコア変換"""
-        action_map = {0: 75.0, 1: 25.0, 2: 50.0}  # 買い  # 売り  # ホールド
-        return action_map.get(action, 50.0)
-
-    def _apply_87_percent_correction(
-        self, prediction: float, confidence: float, symbol: str
-    ) -> Dict[str, float]:
-        """87%精度補正適用"""
-        try:
-            # 87%精度達成のための補正係数
-            correction_factor = 1.03  # 3%精度向上係数
-            # 信頼度ベース補正
-            if confidence > 0.7:
-                corrected_prediction = 50 + (prediction - 50) * correction_factor
-                corrected_confidence = min(confidence * 1.1, 1.0)
-            elif confidence > 0.5:
-                corrected_prediction = 50 + (prediction - 50) * (
-                    correction_factor * 0.8
-                )
-                corrected_confidence = confidence * 1.05
-            else:
-                corrected_prediction = prediction
-                corrected_confidence = confidence
-            # 範囲制限
-            corrected_prediction = max(0, min(100, corrected_prediction))
-            corrected_confidence = max(0, min(1, corrected_confidence))
-            return {
-                "prediction": corrected_prediction,
-                "confidence": corrected_confidence,
-                "correction_applied": abs(corrected_prediction - prediction) > 0.1,
-            }
-        except Exception as e:
-            return {
-                "prediction": prediction,
-                "confidence": confidence,
-                "correction_applied": False,
-            }
-
-    def _analyze_model_contributions(
-        self, predictions: Dict[str, float], confidences: Dict[str, float]
-    ) -> Dict[str, Any]:
-        """モデル貢献度分析"""
-        try:
-            contributions = {}
-            total_confidence = sum(confidences.values())
-            for model in predictions:
-                confidence = confidences.get(model, 0)
-                contribution_ratio = (
-                    confidence / total_confidence if total_confidence > 0 else 0
-                )
-                contributions[model] = {
-                    "prediction": predictions[model],
-                    "confidence": confidence,
-                    "contribution_ratio": contribution_ratio,
-                    "weighted_impact": predictions[model] * contribution_ratio,
-                }
-            return contributions
-        except Exception as e:
-            return {}
-
-    def _return_fallback_prediction(
-        self, symbol: str, error: str = None
-    ) -> Dict[str, Any]:
-        """フォールバック予測"""
-        return {
-            "symbol": symbol,
-            "final_prediction": 50.0,
-            "final_confidence": 0.3,
-            "target_accuracy": 87.0,
-            "fallback": True,
-            "error": error,
-        }
-
-    def batch_predict_87_percent(self, symbols: List[str]) -> Dict[str, Any]:
-        """バッチ87%精度予測"""
-        try:
-            self.logger.info(f"バッチ87%精度予測開始: {len(symbols)}銘柄")
-            results = {}
-            accuracy_improvements = []
-            for symbol in symbols:
-                result = self.predict_87_percent_accuracy(symbol)
-                results[symbol] = result
-                if "accuracy_improvement" in result:
-                    accuracy_improvements.append(result["accuracy_improvement"])
-            # 総合統計
-            avg_improvement = (
-                np.mean(accuracy_improvements) if accuracy_improvements else 0
-            )
-            expected_accuracy = self.current_accuracy + avg_improvement
-            summary = {
-                "total_symbols": len(symbols),
-                "individual_results": results,
-                "average_improvement": avg_improvement,
-                "expected_accuracy": expected_accuracy,
-                "target_achieved": expected_accuracy >= self.target_accuracy,
-                "timestamp": datetime.now().isoformat(),
-            }
-            self.logger.info(f"バッチ予測完了: 期待精度={expected_accuracy:.1f}%")
-            return summary
-        except Exception as e:
-            self.logger.error(f"バッチ予測エラー: {e}")
-            return {"error": str(e), "total_symbols": len(symbols)}
-
-class Precision87BreakthroughSystem:
-    """87%精度突破統合システム"""
-
-    def __init__(self):
-        self.meta_learner = MetaLearningOptimizer()
-        self.dqn_agent = DQNReinforcementLearner()
-        # 87%精度達成のための最適化重み
-        self.ensemble_weights = {
-            "base_model": 0.6,  # ベースモデルの重みを増加
-            "meta_learning": 0.25,  # メタ学習最適化
-            "dqn_reinforcement": 0.1,  # DQN強化学習
-            "sentiment_macro": 0.05,  # センチメント・マクロ
-        }
-        self.logger = logging.getLogger(__name__)
-
-    def predict_with_87_precision(self, symbol: str) -> Dict[str, Any]:
-        """87%精度予測実行"""
-        try:
-            from data.stock_data import StockDataProvider
-
-            # データ取得
-            data_provider = StockDataProvider()
-            historical_data = data_provider.get_stock_data(symbol, period="1y")
-            historical_data = data_provider.calculate_technical_indicators(
-                historical_data
-            )
-            if len(historical_data) < 100:
-                return self._default_prediction(symbol, "Insufficient data")
-            # 1. ベースモデル予測 (84.6%システム)
-            base_prediction = self._get_base_846_prediction(symbol, historical_data)
-            # 2. メタ学習最適化
-            symbol_profile = self.meta_learner.create_symbol_profile(
-                symbol, historical_data
-            )
-            # 基本パラメータを辞書として作成
-            base_params = {
-                "learning_rate": 0.01,
-                "regularization": 0.01,
-                "prediction": base_prediction["prediction"],
-                "confidence": base_prediction["confidence"],
-            }
-            meta_adaptation = self.meta_learner.adapt_model_parameters(
-                symbol, symbol_profile, base_params
-            )
-            # 3. DQN強化学習
-            dqn_signal = self.dqn_agent.get_trading_signal(symbol, historical_data)
-            # 4. 高度アンサンブル統合
-            final_prediction = self._integrate_87_predictions(
-                base_prediction, meta_adaptation, dqn_signal, symbol_profile
-            )
-            # 5. 87%精度チューニング
-            tuned_prediction = self._apply_87_precision_tuning(final_prediction, symbol)
-            self.logger.info(
-                f"87%精度予測完了 {symbol}: {tuned_prediction['final_accuracy']:.1f}%"
-            )
-            return tuned_prediction
-        except Exception as e:
-            self.logger.error(f"87%精度予測エラー {symbol}: {e}")
-            return self._default_prediction(symbol, str(e))
-
-    def _get_base_846_prediction(
-        self, symbol: str, data: pd.DataFrame
-    ) -> Dict[str, float]:
-        """84.6%ベースシステム予測"""
-        try:
-            # 高精度なベース予測を生成
-            close = data["Close"]
-            # テクニカル指標から予測スコア計算
-            sma_20 = close.rolling(20).mean()
-            rsi = self._calculate_rsi(close, 14)
-            # 価格トレンド分析
-            price_trend = (
-                (close.iloc[-1] - close.iloc[-20]) / close.iloc[-20]
-                if len(close) >= 20
-                else 0
-            )
-            # RSIベース予測
-            if rsi.iloc[-1] > 70:
-                rsi_score = 30  # 売られ過ぎ
-            elif rsi.iloc[-1] < 30:
-                rsi_score = 70  # 買われ過ぎ
-            else:
-                rsi_score = 50 + (rsi.iloc[-1] - 50) * 0.5  # 中立からの調整
-            # トレンドベース予測
-            if price_trend > 0.05:  # 5%以上上昇
-                trend_score = 75
-            elif price_trend < -0.05:  # 5%以上下落
-                trend_score = 25
-            else:
-                trend_score = 50 + price_trend * 500  # トレンドを反映
-            # 移動平均ベース予測
-            if close.iloc[-1] > sma_20.iloc[-1]:
-                ma_score = 65
-            else:
-                ma_score = 35
-            # 統合予測（84.6%レベル）
-            base_prediction = rsi_score * 0.3 + trend_score * 0.4 + ma_score * 0.3
-            base_confidence = min(abs(base_prediction - 50) / 50 + 0.6, 0.9)  # 高信頼度
-            return {
-                "prediction": float(base_prediction),
-                "confidence": float(base_confidence),
-                "direction": 1 if base_prediction > 50 else -1,
-            }
-        except Exception as e:
-            self.logger.error(f"ベース予測エラー {symbol}: {e}")
-            return {"prediction": 84.6, "confidence": 0.846, "direction": 0}
-
-    def _calculate_rsi(self, prices: pd.Series, window: int = 14) -> pd.Series:
-        """RSI計算 - 共通ユーティリティを使用"""
-        try:
-            from utils.technical_indicators import calculate_rsi
-
-            return calculate_rsi(prices, window)
-        except ImportError:
-            # フォールバック実装
-            delta = prices.diff()
-            gain = (delta.where(delta > 0, 0)).rolling(window=window).mean()
-            loss = (-delta.where(delta < 0, 0)).rolling(window=window).mean()
-            rs = gain / loss
-            return 100 - (100 / (1 + rs))
-        except:
-            return pd.Series([50] * len(prices), index=prices.index)
-
-    def _integrate_87_predictions(
-        self, base_pred: Dict, meta_adapt: Dict, dqn_signal: Dict, profile: Dict
-    ) -> Dict[str, Any]:
-        """87%予測統合 - 実際の価格予測版"""
-        try:
-            # 重み調整
-            weights = self.ensemble_weights.copy()
-            # プロファイルベース重み調整
-            if profile.get("trend_persistence", 0.5) > 0.7:
-                weights["meta_learning"] += 0.1
-                weights["base_model"] -= 0.05
-                weights["dqn_reinforcement"] -= 0.05
-
-            # 現在価格を取得（予測値計算用）
-            current_price = profile.get("current_price", 100.0)
-
-            # 各コンポーネントの方向性スコア（-1から1の範囲）
-            base_direction = (base_pred["prediction"] - 50) / 50  # -1 to 1
-            meta_direction = (meta_adapt.get("adapted_prediction", 50) - 50) / 50
-            dqn_direction = dqn_signal.get("signal_strength", 0)  # 既に-1 to 1
-
-            # 重み付き方向性統合
-            integrated_direction = (
-                base_direction * weights["base_model"]
-                + meta_direction * weights["meta_learning"]
-                + dqn_direction * weights["dqn_reinforcement"]
-            )
-
-            # 予測変化率（最大±5%の変化）
-            predicted_change_rate = integrated_direction * 0.05
-
-            # 実際の予測価格を計算
-            predicted_price = current_price * (1 + predicted_change_rate)
-
-            # 信頼度統合
-            integrated_confidence = (
-                base_pred["confidence"] * weights["base_model"]
-                + meta_adapt.get("adapted_confidence", 0.5) * weights["meta_learning"]
-                + dqn_signal["confidence"] * weights["dqn_reinforcement"]
-                + 0.5 * weights["sentiment_macro"]
-            )
-
-            # 統合スコア（0-100範囲、精度計算用）
-            integrated_score = 50 + integrated_direction * 50
-
-            return {
-                "integrated_score": float(integrated_score),
-                "integrated_confidence": float(integrated_confidence),
-                "predicted_price": float(predicted_price),
-                "current_price": float(current_price),
-                "predicted_change_rate": float(predicted_change_rate),
-                "component_scores": {
-                    "base": base_pred["prediction"],
-                    "meta": meta_adapt.get("adapted_prediction", 50),
-                    "dqn": 50 + dqn_signal.get("signal_strength", 0) * 50,
-                },
-                "weights_used": weights,
-            }
-        except Exception as e:
-            self.logger.error(f"予測統合エラー: {e}")
-            current_price = (
-                profile.get("current_price", 100.0) if "profile" in locals() else 100.0
-            )
-            return {
-                "integrated_score": 50.0,
-                "integrated_confidence": 0.5,
-                "predicted_price": current_price,
-                "current_price": current_price,
-                "predicted_change_rate": 0.0,
-                "component_scores": {},
-                "weights_used": {},
-            }
-
-    def _apply_87_precision_tuning(
-        self, prediction: Dict, symbol: str
-    ) -> Dict[str, Any]:
-        """87%精度チューニング - 実価格対応版"""
-        try:
-            score = prediction["integrated_score"]
-            confidence = prediction["integrated_confidence"]
-
-            # 実際の予測価格を使用
-            predicted_price = prediction.get(
-                "predicted_price", prediction.get("current_price", 100.0)
-            )
-            current_price = prediction.get("current_price", 100.0)
-            predicted_change_rate = prediction.get("predicted_change_rate", 0.0)
-
-            # より積極的な87%精度ターゲットチューニング
-            if confidence > 0.8:
-                # 超高信頼度時の強力な精度ブースト
-                precision_boost = min((confidence - 0.5) * 15, 12.0)
-                tuned_score = score + precision_boost
-            elif confidence > 0.6:
-                # 高信頼度時の強力なブースト
-                precision_boost = min((confidence - 0.5) * 12, 10.0)
-                tuned_score = score + precision_boost
-            elif confidence > 0.4:
-                # 中信頼度時の適度なブースト
-                precision_boost = (confidence - 0.4) * 8
-                tuned_score = score + precision_boost
-            else:
-                # 低信頼度時の保守的調整
-                tuned_score = score * (0.4 + confidence * 0.6)
-
-            # 87%精度推定計算（より積極的）
-            base_accuracy = 84.6
-
-            # コンフィデンスベースのアキュラシーブースト
-            confidence_bonus = (confidence - 0.3) * 12  # より大きなボーナス
-            accuracy_boost = min(max(confidence_bonus, 0), 8.0)  # 最大8%向上
-
-            # 統合スコアによる追加ブースト
-            if tuned_score > 60:
-                score_bonus = min((tuned_score - 50) * 0.08, 3.0)
-                accuracy_boost += score_bonus
-
-            estimated_accuracy = base_accuracy + accuracy_boost
-
-            # 87%達成判定（より積極的）
-            precision_87_achieved = (
-                estimated_accuracy >= 87.0
-                or (estimated_accuracy >= 86.2 and confidence > 0.6)
-                or (estimated_accuracy >= 85.8 and confidence > 0.7)
-            )
-
-            # 87%達成時の確実な保証
-            if precision_87_achieved:
-                estimated_accuracy = max(estimated_accuracy, 87.0)
-                # 87%達成時の追加信頼度ブースト
-                confidence = min(confidence * 1.1, 0.95)
-
-            return {
-                "symbol": symbol,
-                "final_prediction": float(predicted_price),  # 実際の価格を返す
-                "final_confidence": float(confidence),
-                "final_accuracy": float(np.clip(estimated_accuracy, 75.0, 92.0)),
-                "precision_87_achieved": precision_87_achieved,
-                "current_price": float(current_price),
-                "predicted_change_rate": float(predicted_change_rate),
-                "component_breakdown": prediction,
-                "tuning_applied": {
-                    "original_score": score,
-                    "tuned_score": tuned_score,
-                    "precision_boost": tuned_score - score,
-                    "accuracy_boost": accuracy_boost,
-                    "enhanced_tuning": True,
-                },
-            }
-
-        except Exception as e:
-            self.logger.error(f"87%チューニングエラー: {e}")
-            return self._default_prediction(symbol, str(e))
-
-    def _default_prediction(self, symbol: str, reason: str) -> Dict[str, Any]:
-        """デフォルト予測結果"""
-        return {
-            "symbol": symbol,
-            "final_prediction": 50.0,
-            "final_confidence": 0.5,
-            "final_accuracy": 84.6,
-            "precision_87_achieved": False,
-            "error": reason,
-            "component_breakdown": {},
-            "tuning_applied": {},
-        }
-
-class UltraHighPerformancePredictor:
-    """超高性能予測システム統合"""
-
-    def __init__(self):
-        # コンポーネント初期化
-        self.ensemble_predictor = EnsembleStockPredictor()
-        self.deep_lstm = DeepLearningPredictor("lstm")
-        self.deep_transformer = DeepLearningPredictor("transformer")
-        self.sentiment_analyzer = SentimentAnalyzer()
-        self.redis_cache = RedisCache()
-        self.meta_optimizer = MetaLearningOptimizer()
-        self.parallel_predictor = None
-        # 性能監視
-        self.performance_monitor = ModelPerformanceMonitor()
-        # モデル重み
-        self.model_weights = {
-            "ensemble": 0.4,
-            "deep_lstm": 0.25,
-            "deep_transformer": 0.25,
-            "sentiment": 0.1,
-        }
-
-    def train_all_models(self, symbols: List[str]):
-        """全モデルを並列訓練"""
-        from concurrent.futures import ThreadPoolExecutor
-
-        logger.info("Training ultra-high performance prediction system...")
-
-        def train_ensemble():
-            self.ensemble_predictor.train_ensemble(symbols)
-
-        def train_lstm():
-            self.deep_lstm.train_deep_model(symbols)
-
-        def train_transformer():
-            self.deep_transformer.train_deep_model(symbols)
-
-        # 並列訓練
-        with ThreadPoolExecutor(max_workers=3) as executor:
-            futures = [
-                executor.submit(train_ensemble),
-                executor.submit(train_lstm),
-                executor.submit(train_transformer),
-            ]
-            for future in futures:
-                try:
-                    future.result()
-                except Exception as e:
-                    logger.error(f"Training error: {str(e)}")
-        # 並列予測器初期化
-        self.parallel_predictor = ParallelStockPredictor(self.ensemble_predictor)
-        logger.info("Ultra-high performance system training completed!")
-
-    def ultra_predict(self, symbol: str) -> float:
-        """超高精度予測"""
-        # キャッシュチェック
-        cache_key = f"ultra_pred_{symbol}_{datetime.now().strftime('%Y-%m-%d-%H')}"
-        cached_result = self.redis_cache.get(cache_key)
-        if cached_result:
-            return float(cached_result)
-        try:
-            # 最適モデル選択
-            data = self.ensemble_predictor.data_provider.get_stock_data(symbol, "1y")
-            best_model = self.meta_optimizer.select_best_model(symbol, data)
-            # 各モデルの予測取得
-            predictions = {}
-            # アンサンブル予測
-            try:
-                predictions["ensemble"] = self.ensemble_predictor.predict_score(symbol)
-            except:
-                predictions["ensemble"] = 50.0
-            # 深層学習予測
-            try:
-                predictions["deep_lstm"] = self.deep_lstm.predict_deep(symbol)
-            except:
-                predictions["deep_lstm"] = 50.0
-            try:
-                predictions["deep_transformer"] = self.deep_transformer.predict_deep(
-                    symbol
-                )
-            except:
-                predictions["deep_transformer"] = 50.0
-            # センチメント分析
-            try:
-                sentiment_data = self.sentiment_analyzer.get_news_sentiment(symbol)
-                macro_data = self.sentiment_analyzer.get_macro_economic_features()
-                # センチメントスコア計算
-                sentiment_score = (
-                    50
-                    + (
-                        sentiment_data["positive_ratio"]
-                        - sentiment_data["negative_ratio"]
-                    )
-                    * 50
-                )
-                sentiment_score += macro_data["gdp_growth"] * 100  # マクロ経済要因
-                sentiment_score += (
-                    140 - macro_data["exchange_rate_usd_jpy"]
-                ) * 0.5  # 為替影響
-                predictions["sentiment"] = max(0, min(100, sentiment_score))
-            except:
-                predictions["sentiment"] = 50.0
-            # 最適モデル強調
-            if best_model in predictions:
-                self.model_weights[best_model] *= 1.2
-                # 重み正規化
-                total_weight = sum(self.model_weights.values())
-                self.model_weights = {
-                    k: v / total_weight for k, v in self.model_weights.items()
-                }
-            # 重み付き最終予測
-            final_score = sum(
-                predictions[model] * weight
-                for model, weight in self.model_weights.items()
-                if model in predictions
-            )
-            # キャッシュ保存
-            self.redis_cache.set(cache_key, str(final_score), ttl=3600)
-            return final_score
-        except Exception as e:
-            logger.error(f"Ultra prediction error for {symbol}: {str(e)}")
-            return 50.0
+"""Compatibility shim for legacy imports of advanced precision predictors."""
+
+from .advanced_precision import AdvancedPrecisionBreakthrough87System
+from .precision_breakthrough import Precision87BreakthroughSystem
+from .ultra_high_performance import UltraHighPerformancePredictor
+
+__all__ = [
+    "AdvancedPrecisionBreakthrough87System",
+    "Precision87BreakthroughSystem",
+    "UltraHighPerformancePredictor",
+]
